requirements.txt
# DALI Legal AI MVP - Requirements
# Development of Ambitious Leading Ideas @ Siyada Tech

# Core dependencies
streamlit>=1.28.0
requests>=2.31.0
chromadb>=0.4.15
pandas>=2.0.0

# Web scraping
firecrawl-py>=0.0.8
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Workflow orchestration (optional for MVP)
prefect>=2.14.0

# Text processing
sentence-transformers>=2.2.2
transformers>=4.35.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.0
typing-extensions>=4.8.0

# Development tools (optional)
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0

# Note: Ollama is installed separately via:
# curl -fsSL https://ollama.com/install.sh | sh
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
Jinja2>=3.1.2
python-multipart>=0.0.7
passlib[bcrypt]>=1.7.4
mysql-connector-python>=8.0.0
itsdangerous>=2.0.0

# Enhanced Knowledge Base Dependencies
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
PyMySQL>=1.0.0
pymssql>=2.2.0
plotly>=5.0.0
pandas>=1.5.0

# Web Scraping Dependencies  
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
PyPDF2>=3.0.0
python-docx>=0.8.11
aiohttp>=3.8.0

# FastAPI enhancements (if not already installed)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# For serving static files and templates
jinja2>=3.1.2
python-multipart>=0.0.6

